New version

+++ src/dataset/dataset.py
@@ -40,44 +40,48 @@ class Dataset:
         :param test: percentage of how many of samples should be in test set
         :return:
         """
-        allData = []
-        train_path = os.path.join(self.processed_data, "out")
-        for item in os.listdir(train_path):
-            allData.append(os.path.join(train_path, item))
-        X_train_eval, X_test = train_test_split(
-            allData, test_size=test_per, random_state=20
-        )
-        X_train, X_valid = train_test_split(
-            X_train_eval, test_size=valid_per, random_state=20
-        )
-        if not os.path.exists(os.path.join(self.processed_data, "train")):
-            os.makedirs(os.path.join(self.processed_data, "train"))
-        if not os.path.exists(os.path.join(self.processed_data, "valid")):
-            os.makedirs(os.path.join(self.processed_data, "valid"))
-        if not os.path.exists(os.path.join(self.processed_data, "test")):
-            os.makedirs(os.path.join(self.processed_data, "test"))
+        if not (
+            os.path.exists(os.path.join(self.processed_data, "train"))
+            and os.path.exists(os.path.join(self.processed_data, "test"))
+            and os.path.exists(os.path.join(self.processed_data, "valid"))
+        ):
+            allData = []
+            train_path = os.path.join(self.processed_data, "out")
+            for item in os.listdir(train_path):
+                allData.append(os.path.join(train_path, item))
+            X_train_eval, X_test = train_test_split(
+                allData, test_size=test_per, random_state=20
+            )
+            X_train, X_valid = train_test_split(
+                X_train_eval, test_size=valid_per, random_state=20
+            )
+            if not os.path.exists(os.path.join(self.processed_data, "train")):
+                os.makedirs(os.path.join(self.processed_data, "train"))
+            if not os.path.exists(os.path.join(self.processed_data, "valid")):
+                os.makedirs(os.path.join(self.processed_data, "valid"))
+            if not os.path.exists(os.path.join(self.processed_data, "test")):
+                os.makedirs(os.path.join(self.processed_data, "test"))
 
-        for item in allData:
-            if item in X_train:
-                shutil.move(
-                    item,
-                    os.path.join(self.processed_data, "train"),
-                )
-            elif item in X_valid:
-                shutil.move(
-                    item,
-                    os.path.join(self.processed_data, "valid"),
-                )
-            elif item in X_test:
-                shutil.move(
-                    item,
-                    os.path.join(self.processed_data, "test"),
-                )
-        os.remove(os.path.join(self.processed_data, "out"))
+            for item in allData:
+                if item in X_train:
+                    shutil.move(
+                        item,
+                        os.path.join(self.processed_data, "train"),
+                    )
+                elif item in X_valid:
+                    shutil.move(
+                        item,
+                        os.path.join(self.processed_data, "valid"),
+                    )
+                elif item in X_test:
+                    shutil.move(
+                        item,
+                        os.path.join(self.processed_data, "test"),
+                    )
         return self
 
 
 def prepare_dataset(configs):
     dataset = Dataset(configs).load_dataset_from_file()
     dataset.split_train_dev_test()
-    return
+    return
